renamed wilson_flow -> gradient_flow, etc. and changed a bit the structure of the gauge action header files
